move tc to renderers
