+++ tinygrad/runtime/ops_python.py
@@ -8,7 +8,7 @@ from tinygrad.helpers import all_same, getenv, flatten
 from tinygrad.device import Compiled, Compiler, Allocator
 from tinygrad.codegen.uops import UOpGraph, UOps
 from tinygrad.ops import BinaryOps, TernaryOps, exec_alu
-from tinygrad.renderer import Renderer
+from tinygrad.renderer import Renderer, CUDARenderer, MetalRenderer, HIPRenderer
 
 def _load(m, i):
   if i < 0 or i >= len(m): raise IndexError(f"load out of bounds, size is {len(m)} and access is {i}")
@@ -181,9 +181,9 @@ class PythonProgram:
 class PythonRenderer(Renderer):
   device = "PYTHON"
   def __init__(self):
-    if getenv("EMULATE_METAL"): self.device, self.has_tensor_cores = "METAL", True
-    if getenv("EMULATE_HSA"): self.device, self.has_tensor_cores = "HSA", True
-    if getenv("EMULATE_CUDA"): self.device, self.has_tensor_cores = "CUDA", True
+    if getenv("EMULATE_METAL"): self.device, self.tensor_cores = "METAL", MetalRenderer.tesnor_cores
+    if getenv("EMULATE_HSA"): self.device, self.tensor_cores = "HSA", HIPRenderer.tesnor_cores
+    if getenv("EMULATE_CUDA"): self.device, self.tensor_cores = "CUDA", CUDARenderer.tesnor_cores
 
   def render(self, name:str, uops:UOpGraph) -> str:
     lops = [(u.uop, u.dtype, [uops.uops.index(v) for v in u.vin], u.arg) for u in uops]
